Added optional type annotations to output, a bugfix, now using Python 3.5 type annotations
- Added the option —annotate which enables type annotation. Type
annotation generates a header file per element type which contains a
single line with the name and type of each column. The format per
column is name:type or name:type[] if the column contains an array in
at least one row. The recognized types are integer, float, boolean and
string.
- Fixed a bug where all listed elements in the DTD file were required
to be in the XML file, otherwise you would get an error.
- Changed the array separator “|” to no longer include spaces, the
spaces were previously added only for visibility during debugging.
- For fun and learning, added Python 3.5 type annotations.