Added an option to generate the headers and a shell script to directly import the CSV into a neo4j DB. Fixed issues with escaped bibtex in DBLP.
- Added a —neo4j option that makes the headers more Neo4J-specific and
that generates a shell script that can be directly used to import the
data into a new Neo4J database using the neo4j-import tool.
- There is some bibtex code in the DBLP library that was causing
issues: the quotes in the bibtex were escaped and LXML does not
recognize this and reads in the quotes and escape-characters as-is.
When writing to CSV, the CSV library then double-escapes these
characters. When reading in the CSV again, such as into the Neo4J
database, this causes ambiguity and errors. Fixed this by removing
escaping slashes in front of quotes in the input so that no
double-escaping occurs.