Added preliminary implementation of creating relations.
- Added preliminary implementation of generating relationships between
element attributes and their elements (such as the attribute “author”
of the element “article”). This will create a separate node file for
the attribute and a relationship file for the relation between the
node(attribute) and the element.
- Switched from unique ids per file to global unique ids. This
facilitates easy relationship addition, since the relationship can just
point to the unique id, regardless of in what file the node is in.
- This implementation is not completely finished yet. Todo is
integration with header generation, with the neo4j command generation,
with non-neo4j mode (so generic relations instead of neo4j-specific)
and handling the case where an element attribute can have attributes
itself; for example an author can have an attribute “bibtex” or “href”
and these are currently ignored.